Add protocol.py and refactor fetch.py
